\docType{package}
\name{fireSim-package}
\alias{fireSim-package}
\title{Simulate fire using different approaches}
\description{
  This package is a collection of different fire routines
  to simulate fire spread
}
\details{
  \tabular{ll}{ Package: \tab fireSim\cr Type: \tab
  Package\cr Version: \tab 0.1.0\cr Date: \tab
  TODAYSDATE\cr License: \tab GPL (>= 2)\cr LazyLoad: \tab
  yes\cr }
}
\author{
  Rainer M Krug \email{Rainer@krugs.de}
}

